Postgres implementation of the load_seed macro
This relies on two optional variables set in your dbt project to
override the S3 bucket name and optionally provide a prefix.

This assumes you've created a pg-friendly format of the seed data ([this
repo](https://github.com/waymark-care/tuva-postgres-seed-converter) can
be used to generate one). This macro implementation also requires that
you have setup your RDS cluster/instance with a IAM Role that has the
proper privileges to access the S3 bucket where the seed data is stored.